+++ python_packages/jupyter_lsp/jupyter_lsp/serverextension.py
@@ -1,5 +1,4 @@
-""" add language server support to the running jupyter notebook application
-"""
+"""add language server support to the running jupyter notebook application"""
 
 import json
 from pathlib import Path