+++ python_packages/jupyter_lsp/jupyter_lsp/types.py
@@ -1,5 +1,4 @@
-""" API used by spec finders and manager
-"""
+"""API used by spec finders and manager"""
 
 import asyncio
 import enum