+++ scripts/combine.py
@@ -1,5 +1,4 @@
-""" Combine multiple runs of robot into a single report and log
-"""
+"""Combine multiple runs of robot into a single report and log"""
 
 import sys
 from pathlib import Path